Split Pay
For HackMidWest 2014, we created Split Pay, an IOS+Android application that allows a user to create a bill for something, and his/her friends can submit payment for a percentage of the bill they wish to pay. Once enough money is generated for the bill, the app then gives all the money to the user who created this bill. Split Pay will be perfect for situations like throwing a party or wanting something for your birthday that is to pricy for one person to pay for.

I was responsible for creating the IOS version of the application.

I worked mainly as a front-end IOS developer. For the backend I used Parse API to push persistence into the cloud. Simplify API was also included in the IOS version for credit card management purposes.

We were runner up for HackMidWest 2014.